requirements.txt
fastapi
uvicorn
gunicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-pinecone
pinecone-client
pypdf
langchain-text-splitters
boto3
